Ubuntu 18.04: Enable xattr=sa
I was resistant to enabling xattr=sa by default, as it is
Linux-specific.  However, I have now opened this can of worms for
/var/log because of journald.  People who are using ZFS on Ubuntu as
their root filesystem are probably pretty solidly in the Linux camp,
so this is probably safe.  I have given an explanation of the situation
in the notes.