Ubuntu: Move /tmp off root filesystem
The HOWTO now recommends that /tmp be either a separate dataset (mounted
legacy from /etc/fstab just like /var/tmp, for the same reason) or a
tmpfs (mounted via the standard systemd unit).  In either case, /tmp is
mounted nodev,nosuid.  This is the default for the standard systemd
tmp.mount unit, not something we are overridding.