Debian & Ubuntu: Use update-initramfs -u
We are trying to perform an "update", so -u would be the correct
choice.  This apparently works fine in the Debian instructions already,
and I see no reason it shouldn't work in Ubuntu, so I'm making it
consistent with Debian.

Debian only mentioned update-initramfs -c in the troubleshooting
instructions for the arcsas driver, so I'm fixing it there too.