Debian & Ubuntu: Harmonize the 64-bit text
On all versions, this had a mistake, where "computer" was mentioned
twice in a row.  On Debian, 64-bit was "preferred"; compared to Ubuntu,
this seemed to be a softening of it from a requirement (since it is in
the System Requirements) section.  It's also not really independent of
the image; if you're going to use an amd64 Live CD, it is required.  But
the point of mentioning it is to discourage 32-bit.  I have changed the
wording to match the ZFS on Linux FAQ and to link to that entry.