Ubuntu: Use a separate boot pool
This way, we only have to restrict the boot pool's features, not the
root pool's features.  At the moment, that gets us dnodesize=auto.  In
the future, this will be more important, as we will eventually want to
use native encryption.

Unfortunately, since GRUB then can't read the root filesystem, it fails
to determine the root pool name and the root=ZFS=rpool/ROOT/ubuntu
becomes root=ZFS=/ROOT/ubuntu, which fails.  There are a number of ways
to fix this, but I've chosen to manually set root=ZFS=rpool/ROOT/ubuntu
in /etc/default/grub.  This avoids patching anything in GRUB, including
/etc/grub.d/10_linux.  For future releases, hopefully I can get some
patch into Ubuntu's GRUB to address this without needing to set anything
in /etc/default/grub.