Debian: Update to match Ubuntu
The Debian Stretch HOWTO now uses a separate bpool and supports LUKS.
I'm also using the newer ZFS packages from stretch-backports.

I removed mentions of rootdelay.  That used to exist, but I'm not sure
if it did in stretch.  Now with zfs-initramfs from stretch-backports,
it is definitely gone.

The remaining delta with Ubuntu is expected.

I personally tested both unencrypted and LUKS installs.